Ajout de l objet auteur dans article et modification angular pour faire apparaitre le nom de l auteur sur l article

+++ springboot-udev/src/main/java/com/example/demo/pojo/Article.java
@@ -18,5 +18,5 @@ public class Article {
     private Long id;
     private String title;
     private String message;
-    private int author;
+    private long author;
 }
